+++ django-data/image/cryoweb/tests/test_helpers.py
@@ -358,13 +358,13 @@ class UploadCryoweb(
                 ERROR)
 
             self.assertIn(
-                "Test upload failed",
+                "a valid CryoWeb dump file?",
                 self.submission.message)
 
             # check async message called
             message = 'Error'
             notification_message = (
-                "Error in importing data: Test upload failed\nIs "
+                "Error in importing data: Is "
                 "'cryoweb_test_data_only.sql' a valid CryoWeb dump file?"
             )