+++ django-data/image/submissions/tests/test_view_update_submission.py
@@ -11,6 +11,7 @@ from django.urls import resolve, reverse
 
 from common.tests import (
     FormMixinTestCase, OwnerMixinTestCase, InvalidFormMixinTestCase)
+from uid.models import Submission
 
 from ..forms import UpdateSubmissionForm
 from ..views import UpdateSubmissionView
@@ -35,8 +36,8 @@ class UpdateSubmissionViewTest(
         self.assertIsInstance(view.func.view_class(), UpdateSubmissionView)
 
     def test_form_inputs(self):
-        self.assertContains(self.response, '<input', 4)
-        self.assertContains(self.response, '<select', 2)
+        self.assertContains(self.response, '<input', 5)
+        self.assertContains(self.response, '<select', 3)
 
     def test_reload_not_found_status_code(self):
         url = reverse('submissions:update', kwargs={'pk': 99})
@@ -58,9 +59,11 @@ class SuccessfulUpdateSubmissionViewTest(SubmissionDataMixin, TestCase):
         self.data = {
             'title': 'test-edited',
             'description': 'edited-submission',
-            'gene_bank_name': 'Cryoweb',
+            'gene_bank_name': 'Cryoweb-edited',
             'gene_bank_country': 1,
-            'organization': 1
+            'organization': 1,
+            "datasource_type": 0,
+            "datasource_version": "test",
             }
 
         self.url = reverse('submissions:update', kwargs={'pk': 1})
@@ -74,6 +77,44 @@ class SuccessfulUpdateSubmissionViewTest(SubmissionDataMixin, TestCase):
         self.submission.refresh_from_db()
         self.assertEqual(self.submission.title, self.data["title"])
         self.assertEqual(self.submission.description, self.data["description"])
+
+    def test_same_submission(self):
+        """Updating a submission with the same parameters of another one"""
+
+        # ok create a new submission object
+        data = {
+            "title": "test",
+            "project": "IMAGE",
+            "description": "image test data",
+            "gene_bank_name": "Cryoweb",
+            "gene_bank_country_id": 1,
+            "datasource_type": 0,
+            "datasource_version": "test",
+            "organization_id": 1,
+            "uploaded_file": "data_source/cryoweb_test_data_only.sql",
+            "created_at": "2018-04-11T13:44:31.988Z",
+            "updated_at": "2018-04-11T13:44:31.988Z",
+            "status": 0,
+            "message": "waiting for data loading",
+            "owner_id": 1
+        }
+
+        submission = Submission.objects.create(**data)
+
+        # define a new url
+        url = reverse('submissions:update', kwargs={'pk': submission.id})
+
+        response = self.client.post(
+            url,
+            self.data,
+            follow=True)
+
+        # assert status code (no redirect)
+        self.assertEqual(response.status_code, 200)
+
+        # check errors
+        form = response.context.get('form')
+        self.assertGreater(len(form.errors), 0)
 
 
 class InvalidUpdateSubmissionViewTest(