+++ django-data/image/submissions/tasks.py
@@ -9,8 +9,9 @@ Created on Tue Jul  9 16:10:06 2019
 import logging
 
 from django.utils import timezone
+from django.template.defaultfilters import truncatechars
 
-from common.constants import ERROR, NEED_REVISION
+from common.constants import ERROR, NEED_REVISION, EMAIL_MAX_BODY_SIZE
 from common.tasks import NotifyAdminTaskMixin
 from uid.models import Submission
 from validation.helpers import construct_validation_message
@@ -27,7 +28,7 @@ class SubmissionTaskMixin():
     """A mixin to extend Task to support UID Submission objects"""
 
     action = None
-    max_body_size = 5000
+    max_body_size = EMAIL_MAX_BODY_SIZE
 
     def get_uid_submission(self, submission_id):
         """Get a UID Submission instance from an id
@@ -81,10 +82,10 @@ class SubmissionTaskMixin():
 
     def mail_to_owner(self, submission_obj, subject, body):
         # truncate message body if necessary
-        if len(body) > self.max_body_size:
-            body = body[:self.max_body_size] + "...[truncated]"
+        new_body = truncatechars(body, self.max_body_size) + "[truncated]"
 
-        submission_obj.owner.email_user(subject, body)
+        # send mail to user
+        submission_obj.owner.email_user(subject, new_body)
 
     def on_failure(self, exc, task_id, args, kwargs, einfo):
         """Override the default on_failure method"""